Remove the EmployeeAssignment entity and make the JobAssignment entity directly dependent on the Employee entity
