--- codbex-jobs/gen/codbex-jobs/api/JobAssignment/EmployeeAssignmentService.ts
@@ -1,138 +1,0 @@
-import { Controller, Get, Post, Put, Delete, response } from "sdk/http"
-import { Extensions } from "sdk/extensions"
-import { EmployeeAssignmentRepository, EmployeeAssignmentEntityOptions } from "../../dao/JobAssignment/EmployeeAssignmentRepository";
-import { ValidationError } from "../utils/ValidationError";
-import { HttpUtils } from "../utils/HttpUtils";
-
-const validationModules = await Extensions.loadExtensionModules("codbex-jobs-JobAssignment-EmployeeAssignment", ["validate"]);
-
-@Controller
-class EmployeeAssignmentService {
-
-    private readonly repository = new EmployeeAssignmentRepository();
-
-    @Get("/")
-    public getAll(_: any, ctx: any) {
-        try {
-            const options: EmployeeAssignmentEntityOptions = {
-                $limit: ctx.queryParameters["$limit"] ? parseInt(ctx.queryParameters["$limit"]) : undefined,
-                $offset: ctx.queryParameters["$offset"] ? parseInt(ctx.queryParameters["$offset"]) : undefined
-            };
-
-            let JobAssignment = parseInt(ctx.queryParameters.JobAssignment);
-            JobAssignment = isNaN(JobAssignment) ? ctx.queryParameters.JobAssignment : JobAssignment;
-
-            if (JobAssignment !== undefined) {
-                options.$filter = {
-                    equals: {
-                        JobAssignment: JobAssignment
-                    }
-                };
-            }
-
-            return this.repository.findAll(options);
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Post("/")
-    public create(entity: any) {
-        try {
-            this.validateEntity(entity);
-            entity.Id = this.repository.create(entity);
-            response.setHeader("Content-Location", "/services/ts/codbex-jobs/gen/codbex-jobs/api/JobAssignment/EmployeeAssignmentService.ts/" + entity.Id);
-            response.setStatus(response.CREATED);
-            return entity;
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Get("/count")
-    public count() {
-        try {
-            return this.repository.count();
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Post("/count")
-    public countWithFilter(filter: any) {
-        try {
-            return this.repository.count(filter);
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Post("/search")
-    public search(filter: any) {
-        try {
-            return this.repository.findAll(filter);
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Get("/:id")
-    public getById(_: any, ctx: any) {
-        try {
-            const id = parseInt(ctx.pathParameters.id);
-            const entity = this.repository.findById(id);
-            if (entity) {
-                return entity;
-            } else {
-                HttpUtils.sendResponseNotFound("EmployeeAssignment not found");
-            }
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Put("/:id")
-    public update(entity: any, ctx: any) {
-        try {
-            entity.Id = ctx.pathParameters.id;
-            this.validateEntity(entity);
-            this.repository.update(entity);
-            return entity;
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    @Delete("/:id")
-    public deleteById(_: any, ctx: any) {
-        try {
-            const id = ctx.pathParameters.id;
-            const entity = this.repository.findById(id);
-            if (entity) {
-                this.repository.deleteById(id);
-                HttpUtils.sendResponseNoContent();
-            } else {
-                HttpUtils.sendResponseNotFound("EmployeeAssignment not found");
-            }
-        } catch (error: any) {
-            this.handleError(error);
-        }
-    }
-
-    private handleError(error: any) {
-        if (error.name === "ForbiddenError") {
-            HttpUtils.sendForbiddenRequest(error.message);
-        } else if (error.name === "ValidationError") {
-            HttpUtils.sendResponseBadRequest(error.message);
-        } else {
-            HttpUtils.sendInternalServerError(error.message);
-        }
-    }
-
-    private validateEntity(entity: any): void {
-        for (const next of validationModules) {
-            next.validate(entity);
-        }
-    }
-
-}
