+++ codbex-jobs/gen/codbex-jobs/dao/Employees/JobAssignment.extensionpoint
@@ -1,0 +1,4 @@
+{
+    "name": "codbex-jobs-Employees-JobAssignment",
+    "description": "Extension Point for the codbex-jobs-Employees-JobAssignment entity"
+}
