--- codbex-jobs/gen/codbex-jobs/dao/JobAssignment/EmployeeAssignment.extensionpoint
@@ -1,4 +1,0 @@
-{
-    "name": "codbex-jobs-JobAssignment-EmployeeAssignment",
-    "description": "Extension Point for the codbex-jobs-JobAssignment-EmployeeAssignment entity"
-}
