--- codbex-jobs/gen/codbex-jobs/dao/JobAssignment/JobAssignment.extensionpoint
@@ -1,4 +1,0 @@
-{
-    "name": "codbex-jobs-JobAssignment-JobAssignment",
-    "description": "Extension Point for the codbex-jobs-JobAssignment-JobAssignment entity"
-}
