+++ codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/dialog-filter/view.extension
@@ -1,0 +1,5 @@
+{
+    "module": "codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/dialog-filter/view.js",
+    "extensionPoint": "codbex-jobs-dialog-window",
+    "description": "codbex-jobs - Application Dialog Window"
+}
