+++ codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/dialog-window/view.js
@@ -1,0 +1,17 @@
+/*
+ * Generated by Eclipse Dirigible based on model and template.
+ *
+ * Do not modify the content as it may be re-generated again.
+ */
+const viewData = {
+    id: "JobAssignment-details",
+    label: "JobAssignment",
+    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/dialog-window/index.html",
+    perspectiveName: "Employees",
+};
+
+if (typeof exports !== 'undefined') {
+    exports.getDialogWindow = function () {
+        return viewData;
+    }
+}
