+++ codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/index.html
@@ -1,0 +1,95 @@
+<!DOCTYPE HTML>
+<html lang="en" xmlns="http://www.w3.org/1999/xhtml" ng-app="page" ng-controller="PageController">
+
+	<head>
+		<meta charset="utf-8">
+		<meta http-equiv="X-UA-Compatible" content="IE=edge">
+		<meta name="viewport" content="width=device-width, initial-scale=1">
+		<link rel="icon" href="data:;base64,iVBORw0KGgo=" dg-brand-icon />
+		<title dg-brand-title></title>
+		<theme></theme>
+		<script type="text/javascript"
+			src="/services/js/resources-core/services/loader.js?id=application-view-js"></script>
+		<link type="text/css" rel="stylesheet"
+			href="/services/js/resources-core/services/loader.js?id=application-view-css" />
+		<script type="text/javascript" src="controller.js"></script>
+	</head>
+
+	<body class="dg-vbox">
+		<fd-message-page glyph="sap-icon--hint" ng-show="data == null">
+			<fd-message-page-title>Select a record to get a list of it's details.</fd-message-page-title>
+		</fd-message-page>
+
+		<fd-toolbar has-title="true" ng-hide="data == null">
+			<fd-toolbar-title>Items <span fd-object-status text="{{dataCount}}" inverted="true"></fd-toolbar-title>
+			<fd-button compact="true" dg-type="transparent" dg-label="Filter" ng-click="openFilter()"></fd-button>
+			<fd-button ng-repeat="action in pageActions" compact="true" dg-type="transparent" dg-label="{{action.label}}" ng-click="triggerPageAction(action)"></fd-button>
+			<fd-toolbar-spacer></fd-toolbar-spacer>
+			<fd-button compact="true" dg-type="emphasized" dg-label="Create" ng-click="createEntity()"
+				state="{{selectedMainEntityId == null ? 'disabled' : ''}}"></fd-button>
+		</fd-toolbar>
+		<fd-scrollbar class="dg-full-height" ng-hide="data == null">
+			<table fd-table display-mode="compact" inner-borders="top" outer-borders="none">
+				<thead fd-table-header sticky="true">
+					<tr fd-table-row>
+						<th fd-table-header-cell>Number</th>
+						<th fd-table-header-cell>StartDate</th>
+						<th fd-table-header-cell>EndDate</th>
+						<th fd-table-header-cell>Organization</th>
+						<th fd-table-header-cell>Department</th>
+						<th fd-table-header-cell>JobPosition</th>
+						<th fd-table-header-cell></th>
+					</tr>
+				</thead>
+				<tbody fd-table-body>
+					<tr fd-table-row hoverable="true" ng-show="data.length == 0">
+						<td fd-table-cell no-data="true">No data available.</td>
+					</tr>
+					<tr fd-table-row hoverable="true" ng-repeat="next in data"
+						dg-selected="next.Id === selectedEntity.Id" ng-click="selectEntity(next)">
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">{{next.Number}}</td>
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">
+							<fd-input type="date" ng-model="next.StartDate" ng-readonly="true"></fd-input>
+						</td>
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">
+							<fd-input type="date" ng-model="next.EndDate" ng-readonly="true"></fd-input>
+						</td>
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">
+							{{optionsOrganizationValue(next.Organization)}}
+						</td>
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">
+							{{optionsDepartmentValue(next.Department)}}
+						</td>
+						<td fd-table-cell ng-click="openDetails(next)" hoverable="true" activable="true">
+							{{optionsJobPositionValue(next.JobPosition)}}
+						</td>
+						<td fd-table-cell fit-content="true">
+							<fd-popover>
+								<fd-popover-control>
+									<fd-button compact="true" glyph="sap-icon--overflow" dg-type="transparent"
+										aria-label="Table Row Menu Button" ng-click="setTristate()">
+									</fd-button>
+								</fd-popover-control>
+								<fd-popover-body dg-align="bottom-right">
+									<fd-menu aria-label="Table Row Menu" no-backdrop="true" no-shadow="true">
+										<fd-menu-item ng-repeat="action in entityActions" title="{{action.label}}" ng-click="triggerEntityAction(action)"></fd-menu-item>
+										<fd-menu-item title="View Details" ng-click="openDetails(next)"></fd-menu-item>
+										<fd-menu-item title="Edit" ng-click="updateEntity(next)"></fd-menu-item>
+										<fd-menu-item title="Delete" ng-click="deleteEntity(next)"></fd-menu-item>
+									</fd-menu>
+								</fd-popover-body>
+							</fd-popover>
+						</td>
+					</tr>
+				</tbody>
+			</table>
+		</fd-scrollbar>
+
+		<fd-pagination total-items="dataCount" items-per-page="dataLimit"
+			items-per-page-options="[10, 20, 50]" page-change="loadPage(pageNumber)"
+			items-per-page-change="loadPage(pageNumber)" items-per-page-placement="top-start" compact="true"
+			display-total-items="true" ng-hide="dataCount == 0">
+		</fd-pagination>
+	</body>
+
+</html>
