+++ codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/view.extension
@@ -1,0 +1,5 @@
+{
+    "module": "codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/view.js",
+    "extensionPoint": "codbex-employees-view",
+    "description": "codbex-jobs - Application View - Details"
+}
