+++ codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/view.js
@@ -1,0 +1,19 @@
+/*
+ * Generated by Eclipse Dirigible based on model and template.
+ *
+ * Do not modify the content as it may be re-generated again.
+ */
+const viewData = {
+    id: "JobAssignment",
+    label: "JobAssignment",
+    factory: "frame",
+    region: "bottom",
+    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/Employees/Employee/JobAssignment/index.html",
+    perspectiveName: "Employees",
+};
+
+if (typeof exports !== 'undefined') {
+    exports.getView = function () {
+        return viewData;
+    }
+}
