--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/controller.js
@@ -1,256 +1,0 @@
-angular.module('page', ["ideUI", "ideView", "entityApi"])
-	.config(["messageHubProvider", function (messageHubProvider) {
-		messageHubProvider.eventIdPrefix = 'codbex-jobs.JobAssignment.EmployeeAssignment';
-	}])
-	.config(["entityApiProvider", function (entityApiProvider) {
-		entityApiProvider.baseUrl = "/services/ts/codbex-jobs/gen/codbex-jobs/api/JobAssignment/EmployeeAssignmentService.ts";
-	}])
-	.controller('PageController', ['$scope', '$http', 'messageHub', 'entityApi', 'Extensions', function ($scope, $http, messageHub, entityApi, Extensions) {
-		//-----------------Custom Actions-------------------//
-		Extensions.get('dialogWindow', 'codbex-jobs-custom-action').then(function (response) {
-			$scope.pageActions = response.filter(e => e.perspective === "JobAssignment" && e.view === "EmployeeAssignment" && (e.type === "page" || e.type === undefined));
-			$scope.entityActions = response.filter(e => e.perspective === "JobAssignment" && e.view === "EmployeeAssignment" && e.type === "entity");
-		});
-
-		$scope.triggerPageAction = function (action) {
-			messageHub.showDialogWindow(
-				action.id,
-				{},
-				null,
-				true,
-				action
-			);
-		};
-
-		$scope.triggerEntityAction = function (action) {
-			messageHub.showDialogWindow(
-				action.id,
-				{
-					id: $scope.entity.Id
-				},
-				null,
-				true,
-				action
-			);
-		};
-		//-----------------Custom Actions-------------------//
-
-		function resetPagination() {
-			$scope.dataPage = 1;
-			$scope.dataCount = 0;
-			$scope.dataLimit = 10;
-		}
-		resetPagination();
-
-		//-----------------Events-------------------//
-		messageHub.onDidReceiveMessage("codbex-jobs.JobAssignment.JobAssignment.entitySelected", function (msg) {
-			resetPagination();
-			$scope.selectedMainEntityId = msg.data.selectedMainEntityId;
-			$scope.loadPage($scope.dataPage);
-		}, true);
-
-		messageHub.onDidReceiveMessage("codbex-jobs.JobAssignment.JobAssignment.clearDetails", function (msg) {
-			$scope.$apply(function () {
-				resetPagination();
-				$scope.selectedMainEntityId = null;
-				$scope.data = null;
-			});
-		}, true);
-
-		messageHub.onDidReceiveMessage("clearDetails", function (msg) {
-			$scope.$apply(function () {
-				$scope.entity = {};
-				$scope.action = 'select';
-			});
-		});
-
-		messageHub.onDidReceiveMessage("entityCreated", function (msg) {
-			$scope.loadPage($scope.dataPage, $scope.filter);
-		});
-
-		messageHub.onDidReceiveMessage("entityUpdated", function (msg) {
-			$scope.loadPage($scope.dataPage, $scope.filter);
-		});
-
-		messageHub.onDidReceiveMessage("entitySearch", function (msg) {
-			resetPagination();
-			$scope.filter = msg.data.filter;
-			$scope.filterEntity = msg.data.entity;
-			$scope.loadPage($scope.dataPage, $scope.filter);
-		});
-		//-----------------Events-------------------//
-
-		$scope.loadPage = function (pageNumber, filter) {
-			let JobAssignment = $scope.selectedMainEntityId;
-			$scope.dataPage = pageNumber;
-			if (!filter && $scope.filter) {
-				filter = $scope.filter;
-			}
-			if (!filter) {
-				filter = {};
-			}
-			if (!filter.$filter) {
-				filter.$filter = {};
-			}
-			if (!filter.$filter.equals) {
-				filter.$filter.equals = {};
-			}
-			filter.$filter.equals.JobAssignment = JobAssignment;
-			entityApi.count(filter).then(function (response) {
-				if (response.status != 200) {
-					messageHub.showAlertError("EmployeeAssignment", `Unable to count EmployeeAssignment: '${response.message}'`);
-					return;
-				}
-				if (response.data) {
-					$scope.dataCount = response.data;
-				}
-				filter.$offset = (pageNumber - 1) * $scope.dataLimit;
-				filter.$limit = $scope.dataLimit;
-				entityApi.search(filter).then(function (response) {
-					if (response.status != 200) {
-						messageHub.showAlertError("EmployeeAssignment", `Unable to list/filter EmployeeAssignment: '${response.message}'`);
-						return;
-					}
-					$scope.data = response.data;
-				});
-			});
-		};
-
-		$scope.selectEntity = function (entity) {
-			$scope.selectedEntity = entity;
-		};
-
-		$scope.openDetails = function (entity) {
-			$scope.selectedEntity = entity;
-			messageHub.showDialogWindow("EmployeeAssignment-details", {
-				action: "select",
-				entity: entity,
-				optionsEmployee: $scope.optionsEmployee,
-				optionsTeam: $scope.optionsTeam,
-				optionsJobRole: $scope.optionsJobRole,
-			});
-		};
-
-		$scope.openFilter = function (entity) {
-			messageHub.showDialogWindow("EmployeeAssignment-filter", {
-				entity: $scope.filterEntity,
-				optionsEmployee: $scope.optionsEmployee,
-				optionsTeam: $scope.optionsTeam,
-				optionsJobRole: $scope.optionsJobRole,
-			});
-		};
-
-		$scope.createEntity = function () {
-			$scope.selectedEntity = null;
-			messageHub.showDialogWindow("EmployeeAssignment-details", {
-				action: "create",
-				entity: {},
-				selectedMainEntityKey: "JobAssignment",
-				selectedMainEntityId: $scope.selectedMainEntityId,
-				optionsEmployee: $scope.optionsEmployee,
-				optionsTeam: $scope.optionsTeam,
-				optionsJobRole: $scope.optionsJobRole,
-			}, null, false);
-		};
-
-		$scope.updateEntity = function (entity) {
-			messageHub.showDialogWindow("EmployeeAssignment-details", {
-				action: "update",
-				entity: entity,
-				selectedMainEntityKey: "JobAssignment",
-				selectedMainEntityId: $scope.selectedMainEntityId,
-				optionsEmployee: $scope.optionsEmployee,
-				optionsTeam: $scope.optionsTeam,
-				optionsJobRole: $scope.optionsJobRole,
-			}, null, false);
-		};
-
-		$scope.deleteEntity = function (entity) {
-			let id = entity.Id;
-			messageHub.showDialogAsync(
-				'Delete EmployeeAssignment?',
-				`Are you sure you want to delete EmployeeAssignment? This action cannot be undone.`,
-				[{
-					id: "delete-btn-yes",
-					type: "emphasized",
-					label: "Yes",
-				},
-				{
-					id: "delete-btn-no",
-					type: "normal",
-					label: "No",
-				}],
-			).then(function (msg) {
-				if (msg.data === "delete-btn-yes") {
-					entityApi.delete(id).then(function (response) {
-						if (response.status != 204) {
-							messageHub.showAlertError("EmployeeAssignment", `Unable to delete EmployeeAssignment: '${response.message}'`);
-							return;
-						}
-						$scope.loadPage($scope.dataPage, $scope.filter);
-						messageHub.postMessage("clearDetails");
-					});
-				}
-			});
-		};
-
-		//----------------Dropdowns-----------------//
-		$scope.optionsEmployee = [];
-		$scope.optionsTeam = [];
-		$scope.optionsJobRole = [];
-
-
-		$http.get("/services/ts/codbex-employees/gen/codbex-employees/api/Employees/EmployeeService.ts").then(function (response) {
-			$scope.optionsEmployee = response.data.map(e => {
-				return {
-					value: e.Id,
-					text: e.Name
-				}
-			});
-		});
-
-		$http.get("/services/ts/codbex-organizations/gen/codbex-organizations/api/Teams/TeamService.ts").then(function (response) {
-			$scope.optionsTeam = response.data.map(e => {
-				return {
-					value: e.Id,
-					text: e.Name
-				}
-			});
-		});
-
-		$http.get("/services/ts/codbex-companies/gen/codbex-companies/api/Companies/JobRoleService.ts").then(function (response) {
-			$scope.optionsJobRole = response.data.map(e => {
-				return {
-					value: e.Id,
-					text: e.Name
-				}
-			});
-		});
-
-		$scope.optionsEmployeeValue = function (optionKey) {
-			for (let i = 0; i < $scope.optionsEmployee.length; i++) {
-				if ($scope.optionsEmployee[i].value === optionKey) {
-					return $scope.optionsEmployee[i].text;
-				}
-			}
-			return null;
-		};
-		$scope.optionsTeamValue = function (optionKey) {
-			for (let i = 0; i < $scope.optionsTeam.length; i++) {
-				if ($scope.optionsTeam[i].value === optionKey) {
-					return $scope.optionsTeam[i].text;
-				}
-			}
-			return null;
-		};
-		$scope.optionsJobRoleValue = function (optionKey) {
-			for (let i = 0; i < $scope.optionsJobRole.length; i++) {
-				if ($scope.optionsJobRole[i].value === optionKey) {
-					return $scope.optionsJobRole[i].text;
-				}
-			}
-			return null;
-		};
-		//----------------Dropdowns-----------------//
-
-	}]);
