--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-filter/controller.js
@@ -1,77 +1,0 @@
-angular.module('page', ["ideUI", "ideView"])
-	.config(["messageHubProvider", function (messageHubProvider) {
-		messageHubProvider.eventIdPrefix = 'codbex-jobs.JobAssignment.EmployeeAssignment';
-	}])
-	.controller('PageController', ['$scope', 'messageHub', 'ViewParameters', function ($scope, messageHub, ViewParameters) {
-
-		$scope.entity = {};
-		$scope.forms = {
-			details: {},
-		};
-
-		let params = ViewParameters.get();
-		if (Object.keys(params).length) {
-			$scope.entity = params.entity ?? {};
-			$scope.selectedMainEntityKey = params.selectedMainEntityKey;
-			$scope.selectedMainEntityId = params.selectedMainEntityId;
-			$scope.optionsEmployee = params.optionsEmployee;
-			$scope.optionsTeam = params.optionsTeam;
-			$scope.optionsJobRole = params.optionsJobRole;
-		}
-
-		$scope.filter = function () {
-			let entity = $scope.entity;
-			const filter = {
-				$filter: {
-					equals: {
-					},
-					notEquals: {
-					},
-					contains: {
-					},
-					greaterThan: {
-					},
-					greaterThanOrEqual: {
-					},
-					lessThan: {
-					},
-					lessThanOrEqual: {
-					}
-				},
-			};
-			if (entity.Id !== undefined) {
-				filter.$filter.equals.Id = entity.Id;
-			}
-			if (entity.Employee !== undefined) {
-				filter.$filter.equals.Employee = entity.Employee;
-			}
-			if (entity.Team !== undefined) {
-				filter.$filter.equals.Team = entity.Team;
-			}
-			if (entity.JobRole !== undefined) {
-				filter.$filter.equals.JobRole = entity.JobRole;
-			}
-			if (entity.JobAssignment !== undefined) {
-				filter.$filter.equals.JobAssignment = entity.JobAssignment;
-			}
-			messageHub.postMessage("entitySearch", {
-				entity: entity,
-				filter: filter
-			});
-			$scope.cancel();
-		};
-
-		$scope.resetFilter = function () {
-			$scope.entity = {};
-			$scope.filter();
-		};
-
-		$scope.cancel = function () {
-			messageHub.closeDialogWindow("EmployeeAssignment-filter");
-		};
-
-		$scope.clearErrorMessage = function () {
-			$scope.errorMessage = null;
-		};
-
-	}]);
