--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-filter/index.html
@@ -1,92 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml" ng-app="page" ng-controller="PageController">
-
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="data:;base64,iVBORw0KGgo=" dg-brand-icon />
-        <title dg-brand-title></title>
-        <theme></theme>
-        <script type="text/javascript"
-            src="/services/js/resources-core/services/loader.js?id=application-view-js"></script>
-        <link type="text/css" rel="stylesheet"
-            href="/services/js/resources-core/services/loader.js?id=application-view-css" />
-
-        <script type="text/javascript" src="controller.js"></script>
-    </head>
-
-    <body class="dg-vbox">
-        <fd-scrollbar class="dg-full-height">
-            <div class="fd-margin--md fd-message-strip fd-message-strip--error fd-message-strip--dismissible"
-                role="alert" ng-show="errorMessage">
-                <p class="fd-message-strip__text">{{ errorMessage }}</p>
-                <fd-button glyph="sap-icon--decline" compact="true" dg-type="transparent" aria-label="Close" in-msg-strip="true" ng-click="clearErrorMessage()">
-		        </fd-button>
-            </div>
-
-            <fd-fieldset ng-form="forms.details">
-                <fd-form-group dg-header="EmployeeAssignment Filter" name="entityForm" class="fd-container fd-form-layout-grid-container">
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idEmployee" dg-colon="true">Employee</fd-form-label>
-                        </div>
-                        <div class="fd-col fd-col-md--8">
-	                        <fd-combobox-input id="idEmployee"
-	                            name="Employee"
-	                            state="{{ forms.details['Employee'].$valid ? '' : 'error' }}"
-	                            ng-model="entity.Employee"
-	                            dropdown-items="optionsEmployee"
-	                            dg-placeholder="Search Employee ...">
-	                        </fd-combobox-input>
-	                    </div>
-                    </fd-form-item>
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idTeam" dg-colon="true">Team</fd-form-label>
-                        </div>
-                        <div class="fd-col fd-col-md--8">
-	                        <fd-combobox-input id="idTeam"
-	                            name="Team"
-	                            state="{{ forms.details['Team'].$valid ? '' : 'error' }}"
-	                            ng-model="entity.Team"
-	                            dropdown-items="optionsTeam"
-	                            dg-placeholder="Search Team ...">
-	                        </fd-combobox-input>
-	                    </div>
-                    </fd-form-item>
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idJobRole" dg-colon="true">JobRole</fd-form-label>
-                        </div>
-                        <div class="fd-col fd-col-md--8">
-	                        <fd-combobox-input id="idJobRole"
-	                            name="JobRole"
-	                            state="{{ forms.details['JobRole'].$valid ? '' : 'error' }}"
-	                            ng-model="entity.JobRole"
-	                            dropdown-items="optionsJobRole"
-	                            dg-placeholder="Search JobRole ...">
-	                        </fd-combobox-input>
-	                    </div>
-                    </fd-form-item>
-                </fd-form-group>
-            </fd-fieldset>
-        </fd-scrollbar>
-
-        <fd-bar bar-design="footer" compact="true" ng-show="action !== 'select'">
-			<fd-bar-right>
-				<fd-bar-element>
-					<fd-button dg-label="Filter" dg-type="emphasized" ng-click="filter()"></fd-button>
-				</fd-bar-element>
-				<fd-bar-element>
-					<fd-button dg-label="Reset" ng-click="resetFilter()"></fd-button>
-				</fd-bar-element>
-				<fd-bar-element>
-					<fd-button dg-label="Cancel" ng-click="cancel()"></fd-button>
-				</fd-bar-element>
-			</fd-bar-right>
-		</fd-bar>
-
-    </body>
-
-</html>
