--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-filter/view.js
@@ -1,17 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const viewData = {
-    id: "EmployeeAssignment-filter",
-    label: "EmployeeAssignment Filter",
-    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-filter/index.html",
-    perspectiveName: "JobAssignment",
-};
-
-if (typeof exports !== 'undefined') {
-    exports.getDialogWindow = function () {
-        return viewData;
-    }
-}
