--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-window/controller.js
@@ -1,71 +1,0 @@
-angular.module('page', ["ideUI", "ideView", "entityApi"])
-	.config(["messageHubProvider", function (messageHubProvider) {
-		messageHubProvider.eventIdPrefix = 'codbex-jobs.JobAssignment.EmployeeAssignment';
-	}])
-	.config(["entityApiProvider", function (entityApiProvider) {
-		entityApiProvider.baseUrl = "/services/ts/codbex-jobs/gen/codbex-jobs/api/JobAssignment/EmployeeAssignmentService.ts";
-	}])
-	.controller('PageController', ['$scope', 'messageHub', 'ViewParameters', 'entityApi', function ($scope, messageHub, ViewParameters, entityApi) {
-
-		$scope.entity = {};
-		$scope.forms = {
-			details: {},
-		};
-		$scope.formHeaders = {
-			select: "EmployeeAssignment Details",
-			create: "Create EmployeeAssignment",
-			update: "Update EmployeeAssignment"
-		};
-		$scope.action = 'select';
-
-		let params = ViewParameters.get();
-		if (Object.keys(params).length) {
-			$scope.action = params.action;
-			$scope.entity = params.entity;
-			$scope.selectedMainEntityKey = params.selectedMainEntityKey;
-			$scope.selectedMainEntityId = params.selectedMainEntityId;
-			$scope.optionsEmployee = params.optionsEmployee;
-			$scope.optionsTeam = params.optionsTeam;
-			$scope.optionsJobRole = params.optionsJobRole;
-		}
-
-		$scope.create = function () {
-			let entity = $scope.entity;
-			entity[$scope.selectedMainEntityKey] = $scope.selectedMainEntityId;
-			entityApi.create(entity).then(function (response) {
-				if (response.status != 201) {
-					messageHub.showAlertError("EmployeeAssignment", `Unable to create EmployeeAssignment: '${response.message}'`);
-					return;
-				}
-				messageHub.postMessage("entityCreated", response.data);
-				$scope.cancel();
-				messageHub.showAlertSuccess("EmployeeAssignment", "EmployeeAssignment successfully created");
-			});
-		};
-
-		$scope.update = function () {
-			let id = $scope.entity.Id;
-			let entity = $scope.entity;
-			entity[$scope.selectedMainEntityKey] = $scope.selectedMainEntityId;
-			entityApi.update(id, entity).then(function (response) {
-				if (response.status != 200) {
-					messageHub.showAlertError("EmployeeAssignment", `Unable to update EmployeeAssignment: '${response.message}'`);
-					return;
-				}
-				messageHub.postMessage("entityUpdated", response.data);
-				$scope.cancel();
-				messageHub.showAlertSuccess("EmployeeAssignment", "EmployeeAssignment successfully updated");
-			});
-		};
-
-		$scope.serviceEmployee = "/services/ts/codbex-employees/gen/codbex-employees/api/Employees/EmployeeService.ts";
-		$scope.serviceTeam = "/services/ts/codbex-organizations/gen/codbex-organizations/api/Teams/TeamService.ts";
-		$scope.serviceJobRole = "/services/ts/codbex-companies/gen/codbex-companies/api/Companies/JobRoleService.ts";
-
-		$scope.cancel = function () {
-			$scope.entity = {};
-			$scope.action = 'select';
-			messageHub.closeDialogWindow("EmployeeAssignment-details");
-		};
-
-	}]);
