--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/dialog-window/index.html
@@ -1,89 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml" ng-app="page" ng-controller="PageController">
-
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="data:;base64,iVBORw0KGgo=" dg-brand-icon />
-        <title dg-brand-title></title>
-        <theme></theme>
-        <script type="text/javascript"
-            src="/services/js/resources-core/services/loader.js?id=application-view-js"></script>
-        <link type="text/css" rel="stylesheet"
-            href="/services/js/resources-core/services/loader.js?id=application-view-css" />
-        <script type="text/javascript" src="controller.js"></script>
-    </head>
-
-    <body class="dg-vbox">
-        <fd-scrollbar class="dg-full-height">
-            <fd-fieldset ng-form="forms.details">
-                <fd-form-group dg-header="{{formHeaders[action]}}" name="entityForm" class="fd-container fd-form-layout-grid-container">
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idEmployee" dg-required="false" dg-colon="true">Employee</fd-form-label>
-                        </div>
-                        <div class="fd-col  fd-col-md--8 ">
-	                        <fd-combobox-input id="idEmployee"
-	                            name="Employee"
-	                            state="{{ forms.details['Employee'].$valid ? '' : 'error' }}"
-	                            ng-required="false"
-	                            ng-model="entity.Employee"
-	                            dg-disabled="action === 'select'"
-	                            dropdown-items="optionsEmployee"
-	                            dg-placeholder="Search Employee ...">
-	                        </fd-combobox-input>
-	                        <i class="sap-icon--information" role="presentation" onclick="javascript:alert('')"></i>
-	                    </div>
-                    </fd-form-item>
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idTeam" dg-required="false" dg-colon="true">Team</fd-form-label>
-                        </div>
-                        <div class="fd-col  fd-col-md--8 ">
-	                        <fd-combobox-input id="idTeam"
-	                            name="Team"
-	                            state="{{ forms.details['Team'].$valid ? '' : 'error' }}"
-	                            ng-required="false"
-	                            ng-model="entity.Team"
-	                            dg-disabled="action === 'select'"
-	                            dropdown-items="optionsTeam"
-	                            dg-placeholder="Search Team ...">
-	                        </fd-combobox-input>
-	                        <i class="sap-icon--information" role="presentation" onclick="javascript:alert('')"></i>
-	                    </div>
-                    </fd-form-item>
-                    <fd-form-item class="fd-row">
-                    	<div class="fd-col fd-col-md--2">
-                        	<fd-form-label for="idJobRole" dg-required="false" dg-colon="true">JobRole</fd-form-label>
-                        </div>
-                        <div class="fd-col  fd-col-md--8 ">
-	                        <fd-combobox-input id="idJobRole"
-	                            name="JobRole"
-	                            state="{{ forms.details['JobRole'].$valid ? '' : 'error' }}"
-	                            ng-required="false"
-	                            ng-model="entity.JobRole"
-	                            dg-disabled="action === 'select'"
-	                            dropdown-items="optionsJobRole"
-	                            dg-placeholder="Search JobRole ...">
-	                        </fd-combobox-input>
-	                        <i class="sap-icon--information" role="presentation" onclick="javascript:alert('')"></i>
-	                    </div>
-                    </fd-form-item>
-                </fd-form-group>
-            </fd-fieldset>
-        </fd-scrollbar>
-
-		<fd-bar bar-design="footer" compact="true" ng-show="action !== 'select'">
-			<fd-bar-right>
-				<fd-bar-element>
-					<fd-button dg-label="{{action === 'create' ? 'Create' : 'Update'}}" dg-type="emphasized" ng-click="action === 'create' ? create() : update()" state="{{ forms.details.$valid ? '' : 'disabled' }}"></fd-button>
-				</fd-bar-element>
-				<fd-bar-element>
-					<fd-button dg-label="Cancel" ng-click="cancel()"></fd-button>
-				</fd-bar-element>
-			</fd-bar-right>
-		</fd-bar>
-    </body>
-
-</html>
