--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/view.js
@@ -1,19 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const viewData = {
-    id: "EmployeeAssignment",
-    label: "EmployeeAssignment",
-    factory: "frame",
-    region: "bottom",
-    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/EmployeeAssignment/index.html",
-    perspectiveName: "JobAssignment",
-};
-
-if (typeof exports !== 'undefined') {
-    exports.getView = function () {
-        return viewData;
-    }
-}
