--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/index.html
@@ -1,70 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml" ng-app="page" ng-controller="PageController">
-
-	<head>
-		<meta charset="utf-8">
-		<meta http-equiv="X-UA-Compatible" content="IE=edge">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<link rel="icon" href="data:;base64,iVBORw0KGgo=" dg-brand-icon />
-		<title dg-brand-title></title>
-		<theme></theme>
-		<script type="text/javascript"
-			src="/services/js/resources-core/services/loader.js?id=application-view-js"></script>
-		<link type="text/css" rel="stylesheet"
-			href="/services/js/resources-core/services/loader.js?id=application-view-css" />
-		<script type="text/javascript" src="controller.js"></script>
-	</head>
-
-	<body class="dg-vbox">
-		<fd-toolbar has-title="true">
-			<fd-toolbar-title>Items <span fd-object-status text="{{dataCount}}" inverted="true"></fd-toolbar-title>
-			<fd-toolbar-spacer></fd-toolbar-spacer>
-			<fd-button compact="true" dg-type="transparent" glyph="sap-icon--add"
-				state="{{action == 'update'? 'disabled' : ''}}" ng-click="createEntity()"></fd-button>
-			<fd-button compact="true" dg-type="transparent" glyph="sap-icon--edit"
-				state="{{selectedEntity == null ? 'disabled' : ''}}" ng-click="updateEntity()"></fd-button>
-			<fd-button compact="true" dg-type="transparent" glyph="sap-icon--delete"
-				state="{{selectedEntity == null || action == 'update'? 'disabled' : ''}}" ng-click="deleteEntity()">
-			</fd-button>
-			<fd-popover>
-				<fd-popover-control>
-					<fd-button compact="true" glyph="sap-icon--overflow" dg-type="transparent"
-						aria-label="Header Menu Button">
-					</fd-button>
-				</fd-popover-control>
-				<fd-popover-body dg-align="bottom-right">
-					<fd-menu aria-label="Header Menu" no-backdrop="true" no-shadow="true">
-						<fd-menu-item title="Filter" ng-click="openFilter()"></fd-menu-item>
-						<fd-menu-item ng-repeat="action in pageActions" title="{{action.label}}"
-							ng-click="triggerPageAction(action)"></fd-menu-item>
-					</fd-menu>
-				</fd-popover-body>
-			</fd-popover>
-		</fd-toolbar>
-		<fd-scrollbar>
-			<fd-list list-type="navigation-indication" byline="true" state="disabled">
-				<fd-list-item ng-repeat="next in data">
-					<fd-list-link navigation-indicator="true" dg-selected="next.Id === selectedEntity.Id"
-						navigated="next.Id === selectedEntity.Id"
-						ng-click="action == 'select' ? selectEntity(next) : ''">
-						<fd-list-content
-							item-title="{{next.Number}}">
-							<fd-list-byline dg-align="left">
-								{{optionsEmployeeContractValue(next.EmployeeContract)}} |
-								{{optionsOrganizationValue(next.Organization)}} |
-								{{optionsDepartmentValue(next.Department)}} |
-								{{next.HireDate | date: "dd/MMM/yyyy"}} |
-								{{optionsJobPositionValue(next.JobPosition)}} |
-								{{optionsTeamValue(next.Team)}}
-							</fd-list-byline>
-						</fd-list-content>
-					</fd-list-link>
-				</fd-list-item>
-			</fd-list>
-			<fd-list-action-item ng-show="dataPage <= dataPages && dataCount > dataLimit" ng-click="loadPage()">
-				Load More ...
-			</fd-list-action-item>
-		</fd-scrollbar>
-	</body>
-
-</html>
