--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/main-details/view.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/main-details/view.js","extensionPoint":"codbex-jobs-view","description":"codbex-jobs - Application View - Main Details"}
