--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/main-details/view.js
@@ -1,19 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const viewData = {
-    id: "JobAssignment-details",
-    label: "JobAssignment",
-    factory: "frame",
-    region: "center",
-    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/main-details/index.html",
-    perspectiveName: "JobAssignment",
-};
-
-if (typeof exports !== 'undefined') {
-    exports.getView = function () {
-        return viewData;
-    }
-}
