--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/tile-portal.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/tile.js","extensionPoint":"portal-tile","description":"codbex-jobs - Application Tile"}
