--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/tile.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/tile.js","extensionPoint":"codbex-jobs-tile","description":"codbex-jobs - Application Tile"}
