--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/tile.js
@@ -1,19 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-exports.getTile = function () {
-    return {
-        group: "codbex-jobs",
-        name: "JobAssignment",
-        caption: "Manage entity Job Assignment",
-        tooltip: "JobAssignment",
-        type: "PRIMARY",
-        report: "false",
-        icon: "/services/web/resources/unicons/file.svg",
-        location: "/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/index.html",
-        order: "100",
-        groupOrder: "103",
-    };
-};
