--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/view.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/view.js","extensionPoint":"codbex-jobs-view","description":"codbex-jobs - Application View"}
