--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/view.js
@@ -1,19 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const viewData = {
-    id: "JobAssignment",
-    label: "JobAssignment",
-    factory: "frame",
-    region: "left",
-    link: "/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/JobAssignment/index.html",
-    perspectiveName: "JobAssignment",
-};
-
-if (typeof exports !== 'undefined') {
-    exports.getView = function () {
-        return viewData;
-    }
-}
