--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/index.html
@@ -1,79 +1,0 @@
-<!--
-
-    Generated by Eclipse Dirigible based on model and template.
-
-    Do not modify the content as it may be re-generated again.
-
--->
-<!DOCTYPE html>
-<html lang="en" ng-app="app" ng-controller="ApplicationController" xmlns="http://www.w3.org/1999/xhtml">
-
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="data:;base64,iVBORw0KGgo=" dg-brand-icon />
-        <title dg-brand-title></title>
-        <script type="text/javascript" src="perspective.js"></script>
-        <theme></theme>
-        <script type="text/javascript"
-            src="/services/js/resources-core/services/loader.js?id=application-perspective-js">
-        </script>
-        <link type="text/css" rel="stylesheet"
-            href="/services/js/resources-core/services/loader.js?id=application-perspective-css" />
-    </head>
-
-    <body dg-embeddable-perspective>
-        <ide-header menu-ext-id="codbex-jobs-menu"></ide-header>
-        <ide-container ng-if="!state.error && !state.isBusy">
-            <ide-layout views-layout-model="layoutModel"></ide-layout>
-        </ide-container>
-        <fd-busy-indicator-extended ng-if="!state.error && state.isBusy" class="dg-fill-parent" dg-size="l">
-            {{state.busyText}}
-        </fd-busy-indicator-extended>
-        <fd-message-page glyph="sap-icon--error" ng-if="state.error">
-            <fd-message-page-title>Encounterd an error while loading</fd-message-page-title>
-        </fd-message-page>
-        <ide-dialogs></ide-dialogs>
-        <ide-status-bar></ide-status-bar>
-        <script type="text/javascript">
-            angular.module('app', ['ngResource', 'ideLayout', 'ideUI', 'ideExtensions'])
-                .constant('branding', {
-                    name: 'Jobs Management Module',
-                    brand: 'codbex',
-                    brandUrl: 'https://www.codbex.com',
-                    icons: {
-                        faviconIco: '/services/web/ide-branding/images/favicon.ico',
-                        favicon32: '/services/web/ide-branding/images/favicon-32x32.png',
-                        favicon16: '/services/web/ide-branding/images/favicon-16x16.png',
-                    },
-                    logo: '/services/web/ide-branding/images/logo.svg',
-                })
-                .constant('extensionPoint', {
-                    perspectives: 'codbex-jobs',
-                    views: 'codbex-jobs-view',
-                    dialogWindows: 'codbex-jobs-dialog-window'
-                })
-                .controller('ApplicationController', ['$scope', 'messageHub', 'Extensions', function ($scope, messageHub, Extensions) {
-                    $scope.state = {
-                        isBusy: true,
-                        error: false,
-                        busyText: 'Loading...',
-                    };
-                    $scope.layoutModel = {
-                        views: []
-                    };
-                    Extensions.get('view', 'codbex-jobs-view').then(function (response) {
-                        $scope.layoutModel = {
-                            views: response.filter(e => !e.isLaunchpad && e.perspectiveName === 'JobAssignment').map(e => e.id)
-                        };
-                        $scope.state.isBusy = false;
-                    }, function (error) {
-                        console.error(error);
-                        $scope.state.error = true;
-                    });
-                }]);
-        </script>
-    </body>
-
-</html>
