--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/perspective-portal.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/perspective.js","extensionPoint":"portal","description":"codbex-jobs - Perspective - JobAssignment"}
