--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/perspective.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-jobs/gen/codbex-jobs/ui/JobAssignment/perspective.js","extensionPoint":"codbex-jobs","description":"codbex-jobs - Perspective - JobAssignment"}
