--- codbex-jobs/gen/codbex-jobs/ui/JobAssignment/perspective.js
@@ -1,18 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const perspectiveData = {
-	id: 'JobAssignment',
-	name: 'JobAssignment',
-	link: '/services/web/codbex-jobs/gen/codbex-jobs/ui/JobAssignment/index.html',
-	order: 103,
-	icon: '/services/web/resources/unicons/0-plus.svg',
-};
-
-if (typeof exports !== 'undefined') {
-	exports.getPerspective = function () {
-		return perspectiveData;
-	}
-}
